added a separate tab switcher button with the new designs
