+++ app/src/main/res/layout/include_experimental_omnibar_toolbar_mockup.xml
@@ -85,7 +85,7 @@
             android:src="@drawable/ic_fire" />
     </FrameLayout>
 
-    <com.duckduckgo.app.browser.TabSwitcherButton
+    <com.duckduckgo.app.browser.tabswitcher.ProductionTabSwitcherButton
         android:id="@+id/tabsMenuMockup"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
