+++ app/src/main/res/layout/include_omnibar_toolbar_mockup.xml
@@ -82,7 +82,7 @@
             android:src="@drawable/ic_fire" />
     </FrameLayout>
 
-    <com.duckduckgo.app.browser.TabSwitcherButton
+    <com.duckduckgo.app.browser.tabswitcher.ProductionTabSwitcherButton
         android:id="@+id/tabsMenuMockup"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
