+++ app/src/main/res/layout/view_experimental_tab_switcher_button.xml
@@ -1,0 +1,37 @@
+<?xml version="1.0" encoding="utf-8"?><!--
+  ~ Copyright (c) 2018 DuckDuckGo
+  ~
+  ~ Licensed under the Apache License, Version 2.0 (the "License");
+  ~ you may not use this file except in compliance with the License.
+  ~ You may obtain a copy of the License at
+  ~
+  ~     http://www.apache.org/licenses/LICENSE-2.0
+  ~
+  ~ Unless required by applicable law or agreed to in writing, software
+  ~ distributed under the License is distributed on an "AS IS" BASIS,
+  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+  ~ See the License for the specific language governing permissions and
+  ~ limitations under the License.
+  -->
+
+<merge xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    tools:parentTag="android.widget.RelativeLayout">
+
+    <!-- This TextView doesn't migrate to DaxTextView, special case for the Tabs icon  -->
+    <TextView
+        tools:ignore="DeprecatedWidgetInXml"
+        android:id="@+id/tabCount"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:textColor="?attr/daxColorPrimaryIcon"
+        android:textSize="11sp"
+        android:fontFamily="sans-serif-black"
+        android:textStyle="bold"
+        android:gravity="center"
+        android:paddingBottom="3dp"
+        android:paddingStart="3dp"
+        android:background="@drawable/ic_tab_switcher_experiment"
+        tools:text="66" />
+
+</merge>
