+++ app/src/main/res/layout/view_fade_omnibar.xml
@@ -280,7 +280,7 @@
                 android:src="@drawable/ic_fire" />
         </FrameLayout>
 
-        <com.duckduckgo.app.browser.TabSwitcherButton
+        <com.duckduckgo.app.browser.tabswitcher.ExperimentalTabSwitcherButton
             android:id="@+id/tabsMenu"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
