+++ common/common-ui-internal/src/main/java/com/duckduckgo/common/ui/internal/experiments/visual/VisualDesignExperimentView.kt
@@ -60,10 +60,6 @@ class VisualDesignExperimentView @JvmOverloads constructor(
         viewModel.onWarmColorsFlagChanged(isChecked)
     }
 
-    private val iconsToggleListener = CompoundButton.OnCheckedChangeListener { _, isChecked ->
-        viewModel.onIconsFlagChanged(isChecked)
-    }
-
     override fun onAttachedToWindow() {
         AndroidSupportInjection.inject(this)
         super.onAttachedToWindow()
@@ -78,7 +74,6 @@ class VisualDesignExperimentView @JvmOverloads constructor(
     private fun render(viewState: ViewState) {
         binding.experimentalUIMode.quietlySetIsChecked(viewState.isBrowserThemingFeatureEnabled, experimentalUIToggleListener)
         binding.experimentalUIColorPalette.quietlySetIsChecked(viewState.useWarmColors, warmColorsPaletteToggleListener)
-        binding.experimentalUIIcons.quietlySetIsChecked(viewState.experimentalIcons, iconsToggleListener)
         Snackbar.make(binding.root, "Selected theme is ${viewState.selectedTheme}", Snackbar.LENGTH_SHORT).show()
     }
 }
